Rename wasm-opcodecnt to wasm-stats
