+++ docs/doc/spectest-interp.1.html
@@ -90,7 +90,7 @@ Parse test.json and run the spec tests
   ALSO</a></h1>
 <a class="Xr" href="wasm-interp.1.html">wasm-interp(1)</a>,
   <a class="Xr" href="wasm-objdump.1.html">wasm-objdump(1)</a>,
-  <a class="Xr" href="wasm-opcodecnt.1.html">wasm-opcodecnt(1)</a>,
+  <a class="Xr" href="wasm-stats.1.html">wasm-stats(1)</a>,
   <a class="Xr" href="wasm-strip.1.html">wasm-strip(1)</a>,
   <a class="Xr" href="wasm-validate.1.html">wasm-validate(1)</a>,
   <a class="Xr" href="wasm2c.1.html">wasm2c(1)</a>,
