+++ docs/doc/wasm-decompile.1.html
@@ -80,7 +80,7 @@ Parse binary file test.wasm and write text file test.dcmp
 <a class="Xr" href="wasm2wat.1.html">wasm2wat(1)</a>,
   <a class="Xr" href="wasm-interp.1.html">wasm-interp(1)</a>,
   <a class="Xr" href="wasm-objdump.1.html">wasm-objdump(1)</a>,
-  <a class="Xr" href="wasm-opcodecnt.1.html">wasm-opcodecnt(1)</a>,
+  <a class="Xr" href="wasm-stats.1.html">wasm-stats(1)</a>,
   <a class="Xr" href="wasm-strip.1.html">wasm-strip(1)</a>,
   <a class="Xr" href="wasm-validate.1.html">wasm-validate(1)</a>,
   <a class="Xr" href="wasm2c.1.html">wasm2c(1)</a>,
