+++ docs/doc/wasm-interp.1.html
@@ -102,7 +102,7 @@ Parse binary file test.wasm, and type-check it
 <h1 class="Sh" id="SEE_ALSO"><a class="permalink" href="#SEE_ALSO">SEE
   ALSO</a></h1>
 <a class="Xr" href="wasm-objdump.1.html">wasm-objdump(1)</a>,
-  <a class="Xr" href="wasm-opcodecnt.1.html">wasm-opcodecnt(1)</a>,
+  <a class="Xr" href="wasm-stats.1.html">wasm-stats(1)</a>,
   <a class="Xr" href="wasm-strip.1.html">wasm-strip(1)</a>,
   <a class="Xr" href="wasm-validate.1.html">wasm-validate(1)</a>,
   <a class="Xr" href="wasm2c.1.html">wasm2c(1)</a>,
