+++ docs/doc/wasm-validate.1.html
@@ -84,7 +84,7 @@ Validate binary file test.wasm
   ALSO</a></h1>
 <a class="Xr" href="wasm-interp.1.html">wasm-interp(1)</a>,
   <a class="Xr" href="wasm-objdump.1.html">wasm-objdump(1)</a>,
-  <a class="Xr" href="wasm-opcodecnt.1.html">wasm-opcodecnt(1)</a>,
+  <a class="Xr" href="wasm-stats.1.html">wasm-stats(1)</a>,
   <a class="Xr" href="wasm-strip.1.html">wasm-strip(1)</a>,
   <a class="Xr" href="wasm2c.1.html">wasm2c(1)</a>,
   <a class="Xr" href="wasm2wat.1.html">wasm2wat(1)</a>,
