+++ include/wabt/binary-reader-stats.h
