+++ man/wasm-decompile.1
@@ -66,7 +66,7 @@ Parse binary file test.wasm and write text file test.dcmp
 .Sh SEE ALSO
 .Xr wasm-interp 1 ,
 .Xr wasm-objdump 1 ,
-.Xr wasm-opcodecnt 1 ,
+.Xr wasm-stats 1 ,
 .Xr wasm-strip 1 ,
 .Xr wasm-validate 1 ,
 .Xr wasm2c 1 ,
