+++ man/wasm-interp.1
@@ -95,7 +95,7 @@ Parse test.wasm and run all its exported functions, setting the value stack size
 .Sh SEE ALSO
 .Xr wasm-decompile 1 ,
 .Xr wasm-objdump 1 ,
-.Xr wasm-opcodecnt 1 ,
+.Xr wasm-stats 1 ,
 .Xr wasm-strip 1 ,
 .Xr wasm-validate 1 ,
 .Xr wasm2c 1 ,
