+++ man/wasm-objdump.1
@@ -40,7 +40,7 @@ Print section offsets instead of file offsets in code disassembly
 .Sh SEE ALSO
 .Xr wasm-decompile 1 ,
 .Xr wasm-interp 1 ,
-.Xr wasm-opcodecnt 1 ,
+.Xr wasm-stats 1 ,
 .Xr wasm-strip 1 ,
 .Xr wasm-validate 1 ,
 .Xr wasm2c 1 ,
