+++ man/wasm-validate.1
@@ -69,7 +69,7 @@ Validate binary file test.wasm
 .Xr wasm-decompile 1 ,
 .Xr wasm-interp 1 ,
 .Xr wasm-objdump 1 ,
-.Xr wasm-opcodecnt 1 ,
+.Xr wasm-stats 1 ,
 .Xr wasm-strip 1 ,
 .Xr wasm2c 1 ,
 .Xr wasm2wat 1 ,
