+++ man/wasm2c.1
@@ -75,7 +75,7 @@ Parse test.wasm, write test.c and test.h, but ignore the debug names, if any
 .Xr wasm-decompile 1 ,
 .Xr wasm-interp 1 ,
 .Xr wasm-objdump 1 ,
-.Xr wasm-opcodecnt 1 ,
+.Xr wasm-stats 1 ,
 .Xr wasm-strip 1 ,
 .Xr wasm-validate 1 ,
 .Xr wasm2wat 1 ,
