+++ man/wasm2wat.1
@@ -85,7 +85,7 @@ Parse test.wasm, write test.wat, but ignore the debug names, if any
 .Xr wasm-decompile 1 ,
 .Xr wasm-interp 1 ,
 .Xr wasm-objdump 1 ,
-.Xr wasm-opcodecnt 1 ,
+.Xr wasm-stats 1 ,
 .Xr wasm-strip 1 ,
 .Xr wasm-validate 1 ,
 .Xr wasm2c 1 ,
