+++ man/wast2json.1
@@ -78,7 +78,7 @@ Modules are written to spec-test.0.wasm, spec-test.1.wasm, etc.
 .Xr wasm-decompile 1 ,
 .Xr wasm-interp 1 ,
 .Xr wasm-objdump 1 ,
-.Xr wasm-opcodecnt 1 ,
+.Xr wasm-stats 1 ,
 .Xr wasm-strip 1 ,
 .Xr wasm-validate 1 ,
 .Xr wasm2c 1 ,
