+++ man/wat2wasm.1
@@ -87,7 +87,7 @@ Parse spec-test.wast, and write verbose output to stdout (including the meaning 
 .Xr wasm-decompile 1 ,
 .Xr wasm-interp 1 ,
 .Xr wasm-objdump 1 ,
-.Xr wasm-opcodecnt 1 ,
+.Xr wasm-stats 1 ,
 .Xr wasm-strip 1 ,
 .Xr wasm-validate 1 ,
 .Xr wasm2c 1 ,
