+++ src/binary-reader-stats.cc
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "wabt/binary-reader-opcnt.h"
+#include "wabt/binary-reader-stats.h"
 
 #include <cassert>
 #include <cinttypes>
