+++ test/stats/cutoff.txt
@@ -1,4 +1,4 @@
-;;; TOOL: run-opcodecnt
+;;; TOOL: run-stats
 ;;; ARGS: --cutoff 3
 (module
   (func
