+++ test/stats/immediates.txt
@@ -1,4 +1,4 @@
-;;; TOOL: run-opcodecnt
+;;; TOOL: run-stats
 (module
   (memory 1)
   (func
